Fix a bug where a long was dereferenced from an int pointer.
git-svn-id: https://warzone2100.svn.sourceforge.net/svnroot/warzone2100/trunk@8584 4a71c877-e1ca-e34f-864e-861f7616d084

+++ lib/ivis_common/tex.h
@@ -37,7 +37,7 @@
 typedef struct
 {
 	char name[iV_TEXNAME_MAX];
-	unsigned int id;
+	unsigned long id;
 } iTexPage;
 
 //*************************************************************************
